the rendering engine (pixi.js) is all set up and functional. the platform engine is pretty much done locally (z to jump). also managed to link nodejs to share part of the code with the client (for server simulation purpose) , but met with some complications
1. need a global at the end of the js file for those included for god knows what reason
2. need another constants file, meaning if we change one constant, we have to change the one for the server.
   i'll see if there is another way.

nonetheless, it manages to run pretty well.

currently you can open 2 clients to play with. i've only coded for 2 players for now.
couple techniques used to implement the networking stuff but not working super well, still have a lot of work to make sure it's smooth.

code is quite messy at the moment, i'll try to clean it up later on.