quite a few new things - camera is up - you can now and get hurt and die (press 'H' to get hurt until you die and respawn) - simple GUI is up indicating life and HP - sound system set up with some random music and jumping sound - can now support more than potentially infinite players - can now create all character types, but most of the animations are not in yet
@Priscilla - not sure how you intend to link the lobby to the game, but i added a new message to create new players.
here's how it flows tentatively -
in index.html, a character is created that belongs to the client right at the start
in index.html, a listener is created to get 'newPlayer' message, which contains the PlayerID (same as socket.id). a new character is created for the client. It also contains the character type.
in server.js, whenever a player joins, 'newPlayer' messages are broadcasted to existing players and to the new player.
feel free to tell me what you need changed to help integrate with the lobby.
on another note, SaveHer.html seems to lack some missing files (like the logo), i can't seem to run it.

@XianQun - i'm not sure if you realise but i created the rectangle class to facilitate collision detection which you may find useful in doing the projectiles. You can of course do your own if you like.
You can actually see all of the rectangles (some hidden behind the actual object) by going to constants.js and setting DEBUGMODE to true.
to make things easier (i hope), you only need 2 functions when interacting with a character,

character.Hurt(dmg) - just does damage to the character, will return false if the damage fails to be dealt (eg: when just spawned, then the guy is invincible for a few seconds). the character will also die when its HP reaches 0 so you don't have to worry about that.
character.isColliding(rect) - determine if the rectangle obj (eg: of your bullet) hits the character. You can reference initDetectors() part in character.js to see how a rectangle is constructed.

All of those are done locally, so it's up to you how to handle the communication (aka the tough part :S). Feel free to edit any parts of my code.

@ZiXian - the sound system is set up so you can start playing with sounds. you can have a look at SoundManager.js to see how it works. I'm using this - http://www.html5rocks.com/en/tutorials/webaudio/intro/
it seems to load the sound every time it's called though, which may be slow. so you can try experimenting with other libraries if you wish.